refactor: UI 로직으로 분리
