+++ src/js/view/currentAmount.js
@@ -1,0 +1,5 @@
+import { $ } from '../utils/querySelector.js';
+
+export const showCurrentAmount = (targetId, currentAmount) => {
+  $(targetId).innerHTML = currentAmount;
+};
